+++ src/main/kotlin/me/nalab/api/survey/domain/feedback/ChoiceFormQuestionFeedback.kt
@@ -1,4 +1,4 @@
-package domain.feedback
+package me.nalab.api.survey.domain.feedback
 
 import javax.persistence.*
 
@@ -7,11 +7,11 @@ class ChoiceFormQuestionFeedback(
     id: Long,
     formQuestionId: Long,
     isRead: Boolean = false,
-    bookmark: domain.feedback.Bookmark = domain.feedback.Bookmark.Companion.impossible(),
-    feedback: domain.feedback.Feedback,
+    bookmark: Bookmark = Bookmark.impossible(),
+    feedback: Feedback,
 
     @ElementCollection
     @Column(name = "selects")
     @CollectionTable(name = "selects", joinColumns = [JoinColumn(name = "form_feedback_id")])
     private val selectedChoiceIds: MutableSet<Long>,
-) : domain.feedback.FormQuestionFeedbackable(id, formQuestionId, isRead, bookmark, feedback)
+) : FormQuestionFeedbackable(id, formQuestionId, isRead, bookmark, feedback)
