+++ src/main/kotlin/me/nalab/api/survey/domain/survey/ChoiceFormQuestion.kt
@@ -1,6 +1,6 @@
-package domain.survey
+package me.nalab.api.survey.domain.survey
 
-import me.nalab.survey.domain.survey.exception.DuplicatedOrderException
+import me.nalab.api.survey.domain.survey.exception.DuplicatedOrderException
 import javax.persistence.*
 
 @Entity
