+++ src/main/kotlin/me/nalab/api/survey/domain/survey/ChoiceFormQuestionType.kt
@@ -1,4 +1,4 @@
-package domain.survey
+package me.nalab.api.survey.domain.survey
 
 enum class ChoiceFormQuestionType {
     TENDENCY,
