+++ src/main/kotlin/me/nalab/api/survey/domain/survey/SurveyRepository.kt
@@ -1,4 +1,4 @@
-package domain.survey
+package me.nalab.api.survey.domain.survey
 
 import org.springframework.data.jpa.repository.JpaRepository
